Delegate several common tasks to included builds
This is a bit of a workaround until the following improvements are made
to Gradle's composite build support (taken from
https://docs.gradle.org/current/userguide/composite_builds.html):

> Improvements we have planned for upcoming releases include:
>
> The ability to target a task or tasks in an included build directly from
the command line. We are currently exploring syntax options for allowing
this functionality, which will remove many cases where a delegating task
is required in the composite.

So for now, we need to create an entry here in the top-level
build.gradle file for each task we want to run across all included
builds.

With these changes, one can run `./gradlew build` at the top level, and
the same command will be delegated to all included builds.

But one cannot run, for example `./gradlew compileJava`, because it
doesn't have an explicit task declaration here (yet).

Note that for finer grained, control, one can alwys fall back to `mr`
(see the README in the root of this repository) to run a given command
in each repository, e.g. `mr run ./gradlew compileJava`.

See bisq-network/dao#41